Improve cpu usage with CUDA.
Switch to fork of mumax/3/cuda/cu that adds calls to the CUDA runtime
api.  ccminer uses the runtime api and so by using that we can use the
same methods to get similar cpu performance.

This also removes the use of CUDA contexts complete.

There are still a few calls (for getting device count, version, etc)
that use the older driver api but you are allowed to mix and match and
those can be converted at a later date if desired.

Closes #86