cuda: Perform autocalibration of the grid size.
The size of the job sent to the GPU is gridsize * threadcount. The
thread count can be roughly predetermined, based on the properties of
the kernel and the currently available device families. This is
currently set by default as 256.

Autocalibration is done in a similar fashion to the OpenCL one, by
determining the size where each job takes about 500ms.