cuda: Switch to cuda_builder.go instead of Makefile.
This switches the method for generating compiled versions of the
CUDA-enabled Blake3 kernel from the existing GNUMakefile to use a go
generate based generator.

This is necessary to properly support Windows builds, removes the direct
need for an external dependency (make) and makes gominer follow a more
idiomatic Go method for building its releases.