cuda: Support CUDA toolkit versions older than 10.
This commit adds a conditional macro to the Blake3 CUDA kernel so that
on older CUDA Toolkit versions a shift-based rotation is used instead of
the intrinsic.

According to the CUDA Toolkit documentation[1], the __funnelshift_rc
integer intrinsic was introduced on CUDA version 10. Prior to this
commit, the CUDA kernel of gominer could not be built when using
Toolkit versions older than 10.

Note that although this makes it possible to build the kenel on older
Toolkit versions, performance is significantly degraded when compared to
the modern, intrinsic-based versions.

[1]: https://docs.nvidia.com/cuda/archive/10.0/cuda-math-api/group__CUDA__MATH__INTRINSIC__INT.html