cuda: Also use GPU arch for intrinsic determination.
Newer CUDA Toolkit versions can still target older GPU architectures for
kernel compilation. Therefore, the target architecture also needs to be
taken into account when determining whether to use the intrinsic or not
for the ROTR macro of the kernel.

This adds a check to only use the intrinsic for architectures greater
than compute_30.