cuda: Automatically determine GPU architecture.
When compiling the CUDA kernel, it is important to specify the GPU
architecture to use in order to use the best possible GPU features,
namely the intrinsic used for the ROTR macro.

Different CUDA Toolkit versions support different architecture
selectors, therefore this commit adds automatic GPU architecture
determination based on the installed Toolkit version. Versions greater
than 11.5 use the 'all' selector for architecture, which is the broadest
available. Older versions default to using the 'compute_50'
architecture, which provides support for the required features.

The automatic selection can be overridden by specifying an environment
variable GOMINER_CUDA_GPU_ARCH when generating the kernel with go
generate which allows testing different configurations.