Perform initial getwork when solo mining.
This ensure work is immediately available when solo mining by making an
initial call to getwork during the initial setup of the miner.

It entails refactoring the work preparation logic from the OnWork
handler into a separate function that is called with the result of the
initial getwork call as well as all future OnWork notifications.